reviews: Rename TemplateKey and TemplateValue, create toml mod, fix errors
