+++ objects/src/accounts/mod.rs
@@ -25,8 +25,8 @@ pub use code::{procedure::AccountProcedureInfo, AccountCode};
 
 mod component;
 pub use component::{
-    AccountComponent, AccountComponentTemplate, ComponentMetadata, InitStorageData, StorageEntry,
-    TemplateKey, TemplateValue,
+    AccountComponent, AccountComponentMetadata, AccountComponentTemplate, InitStorageData,
+    StorageEntry, StoragePlaceholder, StorageValue,
 };
 
 pub mod delta;
@@ -463,8 +463,7 @@ mod tests {
                 Digest::new([Felt::new(105), Felt::new(106), Felt::new(107), Felt::new(108)]),
                 [Felt::new(5_u64), Felt::new(6_u64), Felt::new(7_u64), Felt::new(8_u64)],
             ),
-        ])
-        .unwrap();
+        ]);
         let storage_slot_map = StorageSlot::Map(storage_map.clone());
 
         let mut account = build_account(
