+++ objects/src/errors.rs
@@ -16,7 +16,7 @@ use super::{
     MAX_OUTPUT_NOTES_PER_BATCH, MAX_OUTPUT_NOTES_PER_BLOCK,
 };
 use crate::{
-    accounts::{AccountCode, AccountIdPrefix, AccountStorage, AccountType},
+    accounts::{AccountCode, AccountIdPrefix, AccountStorage, AccountType, StoragePlaceholder},
     block::block_num_from_epoch,
     notes::{NoteAssets, NoteExecutionHint, NoteTag, NoteType, Nullifier},
     ACCOUNT_UPDATE_MAX_SIZE, MAX_INPUTS_PER_NOTE, MAX_INPUT_NOTES_PER_TX, MAX_OUTPUT_NOTES_PER_TX,
@@ -27,8 +27,6 @@ use crate::{
 
 #[derive(Debug, Error)]
 pub enum AccountComponentTemplateError {
-    #[error("error creating AccountComponent")]
-    AccountComponentError(#[source] AccountError),
     #[cfg(feature = "std")]
     #[error("error trying to deserialize from toml")]
     DeserializationError(#[source] toml::de::Error),
@@ -44,10 +42,8 @@ pub enum AccountComponentTemplateError {
     MetadataDeserializationError(String),
     #[error("component storage slots are not contiguous ({0} is followed by {1})")]
     NonContiguousSlots(u8, u8),
-    #[error("error creating storage map")]
-    StorageMapError(#[source] AccountError),
-    #[error("template value ({0}) was not provided in the map")]
-    TemplateValueNotProvided(String),
+    #[error("template value for placeholder ({0}) was not provided in the map")]
+    PlaceholderValueNotProvided(StoragePlaceholder),
 }
 
 // ACCOUNT ERROR
@@ -55,11 +51,6 @@ pub enum AccountComponentTemplateError {
 
 #[derive(Debug, Error)]
 pub enum AccountError {
-    #[error("failed to assemble account component:\n{}", PrintDiagnostic::new(.0))]
-    AccountComponentAssemblyError(Report),
-    // TODO: Use MastForestError once it implements Error in no-std.
-    #[error("failed to merge account code: {0}")]
-    AccountComponentMergeError(String),
     #[error("failed to deserialize account code")]
     AccountCodeDeserializationError(#[source] DeserializationError),
     #[error("account code does not contain procedures but must contain at least one procedure")]
@@ -72,6 +63,13 @@ pub enum AccountError {
     AccountCodeProcedureStorageSizeTooLarge(Digest, Felt),
     #[error("account procedure {0}'s final two elements must be Felt::ZERO")]
     AccountCodeProcedureInvalidPadding(Digest),
+    #[error("failed to assemble account component:\n{}", PrintDiagnostic::new(.0))]
+    // TODO: Use MastForestError once it implements Error in no-std.
+    AccountComponentAssemblyError(Report),
+    #[error("failed to merge account code: {0}")]
+    AccountComponentMergeError(String),
+    #[error("failed to create account component")]
+    AccountComponentTemplateInstantiationError(#[source] AccountComponentTemplateError),
     #[error("failed to convert bytes into account id field element")]
     AccountIdInvalidFieldElement(#[source] DeserializationError),
     #[error("failed to update asset vault")]
